+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* New fields for testing: Real Time Attendance, Lessons Date
+
 ### 1.4.0: April 14th, 2021
 * Gravity Forms: Add WooCommerce styling to message content
 * Gravity Forms: Customize email notification (headers, message, to, subject) to use WooCommerce CSS styling